pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "fpdf>=1.7.2",
    "google-adk>=1.0.0",
    "google-genai>=1.16.1",
    "graphviz>=0.20.3",
    "litellm>=1.71.1",
    "nltk>=3.9.1",
    "pandas>=2.3.0",
    "pyvis>=0.3.2",
]
